requirements.txt
streamlit
openai
requests
bs4
anthropic
google-generativeai